backend/requirements.txt
# ── Web 框架 ──────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9          # 文件上传 / Form 支持

# ── 数据校验与配置 ─────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ── ORM 与数据库迁移 ───────────────────────────────────────────────────────────
sqlalchemy>=2.0.30
alembic>=1.13.1
pymysql>=1.1.1
cryptography>=42.0.0             # PyMySQL SSL 依赖

# ── 缓存与消息队列 ─────────────────────────────────────────────────────────────
redis>=5.0.4
celery[redis]>=5.3.6

# ── 安全与认证 ─────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0  # JWT 签发与校验
passlib[bcrypt]>=1.7.4            # 密码哈希
bcrypt>=3.2.0,<4.0                # passlib 与 bcrypt 4.x 存在兼容问题，固定 3.x

# ── IoT 接入适配 ───────────────────────────────────────────────────────────────
paho-mqtt>=2.1.0                  # MQTT 南向接入
httpx>=0.27.0                     # 异步 HTTP（OPC-UA REST 网关 / 外部接口调用）

# ── 工具库 ─────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
tenacity>=8.3.0                   # 重试策略（IoT 重连、仿真任务容错）
